refactor: ♻️ rename integration folder to match domain
- rename custom_components/luxtronik/ to custom_components/luxtronik2/ to match manifest domain
- update CI workflow paths in linter.yml and test.yaml
- update codespell skip path in pyproject.toml
- fix import sort order in model.py (luxtronik now third-party after rename)
- update disabled test imports to custom_components.luxtronik2
