+++ .github/workflows/test.yaml
@@ -12,26 +12,25 @@ jobs:
         python-version: ["3.13", "3.14"]
 
     steps:
-    - name: Checkout code
-      uses: actions/checkout@v6
+      - name: Checkout code
+        uses: actions/checkout@v6
 
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v6
-      with:
-        python-version: ${{ matrix.python-version }}
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v6
+        with:
+          python-version: ${{ matrix.python-version }}
 
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install -r tests/requirements-dev.txt
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r tests/requirements-dev.txt
 
-    - name: Run tests with coverage
-      run: |
-        pytest --cov=custom_components.luxtronik --cov-report=xml --cov-report=term tests/
+      - name: Run tests with coverage
+        run: |
+          pytest --cov=custom_components.luxtronik2 --cov-report=xml --cov-report=term tests/
 
-    - name: Upload coverage report
-      uses: actions/upload-artifact@v7
-      with:
-        name: coverage-report
-        path: coverage.xml
-
+      - name: Upload coverage report
+        uses: actions/upload-artifact@v7
+        with:
+          name: coverage-report
+          path: coverage.xml
