+++ custom_components/luxtronik2/binary_sensor.py
