+++ custom_components/luxtronik2/binary_sensor_entities_predefined.py
