+++ custom_components/luxtronik2/const.py
