+++ custom_components/luxtronik2/date_entities_predefined.py
