+++ custom_components/luxtronik2/diagnostics.py
