+++ custom_components/luxtronik2/manifest.json
