+++ custom_components/luxtronik2/model.py
@@ -27,9 +27,8 @@ from homeassistant.components.water_heater import (
 from homeassistant.const import Platform, UnitOfTemperature
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.typing import StateType
-from packaging.version import Version
-
 from luxtronik import Calculations, Parameters, Visibilities
+from packaging.version import Version
 
 from .const import (
     UPDATE_INTERVAL_VERY_SLOW,
