+++ custom_components/luxtronik2/schema_helper.py
