+++ custom_components/luxtronik2/sensor_entities_predefined.py
