+++ custom_components/luxtronik2/switch_entities_predefined.py
