+++ custom_components/luxtronik2/translations/cs.json
