+++ custom_components/luxtronik2/translations/de.json
