+++ custom_components/luxtronik2/translations/en.json
