+++ custom_components/luxtronik2/translations/pl.json
