+++ tests/disabled_test_options_flow.py
@@ -2,8 +2,8 @@ from homeassistant import data_entry_flow
 from homeassistant.core import HomeAssistant
 import pytest
 
-from custom_components.luxtronik.config_flow import LuxtronikOptionsFlowHandler
-from custom_components.luxtronik.const import (
+from custom_components.luxtronik2.config_flow import LuxtronikOptionsFlowHandler
+from custom_components.luxtronik2.const import (
     CONF_HA_SENSOR_INDOOR_TEMPERATURE,
     CONF_HOST,
     CONF_PORT,
